Move window creation and font rendering to libuser
- Move clock::vbe to libuser::window and libuser::terminal
- Remove clock::logger
