+++ clock/Cargo.toml
@@ -5,11 +5,3 @@ authors = ["roblabla <unfiltered@roblab.la>", "orycterope <tvermeilh@gmail.com>"
 
 [dependencies]
 kfs-libuser = { path = "../libuser" }
-kfs-libutils = { path = "../libutils" }
-font-rs = { git = "https://github.com/orycterope/font-rs" }
-spin = "0.4"
-hashmap_core = "0.1.9"
-
-[dependencies.lazy_static]
-features = ["spin_no_std"]
-version = "1.0.0"
