+++ libuser/Cargo.toml
@@ -11,6 +11,8 @@ spin = "0.4"
 kfs-libutils = { path = "../libutils" }
 kfs-libkern = { path = "../libkern" }
 failure = { version = "0.1", default-features = false, features = ["derive"] }
+font-rs = { git = "https://github.com/orycterope/font-rs" }
+hashmap_core = "0.1.9"
 
 [dependencies.byteorder]
 default-features = false
