+++ libuser/src/lib.rs
@@ -21,6 +21,8 @@ extern crate kfs_libutils;
 extern crate kfs_libkern;
 #[macro_use]
 extern crate failure;
+extern crate font_rs;
+extern crate hashmap_core;
 
 pub mod syscalls;
 pub mod io;
@@ -30,6 +32,8 @@ pub mod sm;
 pub mod vi;
 pub mod error;
 pub mod allocator;
+pub mod terminal;
+pub mod window;
 
 use kfs_libutils as utils;
 use error::{Error, LibuserError};
